Refactor: Rename ui/team package to ui/teams
Renamed the `ui/team` directory to `ui/teams` and updated all package declarations and imports throughout the codebase to reflect the change. This improves consistency and clarity in the project structure.
