+++ app/src/main/java/org/ole/planet/myplanet/ui/community/CommunityPagerAdapter.kt
@@ -9,7 +9,7 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.ui.enterprises.EnterprisesFinancesFragment
 import org.ole.planet.myplanet.ui.enterprises.EnterprisesReportsFragment
 import org.ole.planet.myplanet.ui.news.NewsFragment
-import org.ole.planet.myplanet.ui.team.TeamCalendarFragment
+import org.ole.planet.myplanet.ui.teams.TeamCalendarFragment
 
 class CommunityPagerAdapter(private val fm: FragmentActivity, private val id: String, private var fromLogin: Boolean, val settings: SharedPreferences) : FragmentStateAdapter(fm) {
     override fun createFragment(position: Int): Fragment {
