+++ app/src/main/java/org/ole/planet/myplanet/ui/teams/TeamDetailFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.team
+package org.ole.planet.myplanet.ui.teams
 
 import android.content.DialogInterface
 import android.os.Bundle
@@ -31,21 +31,21 @@ import org.ole.planet.myplanet.service.sync.SyncManager
 import org.ole.planet.myplanet.service.UserProfileDbHandler
 import org.ole.planet.myplanet.service.sync.RealtimeSyncCoordinator
 import org.ole.planet.myplanet.service.sync.ServerUrlMapper
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.ApplicantsPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.CalendarPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.ChatPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.CoursesPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.DocumentsPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.FinancesPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.JoinRequestsPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.MembersPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.MissionPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.PlanPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.ReportsPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.ResourcesPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.SurveyPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.TasksPage
-import org.ole.planet.myplanet.ui.team.TeamPageConfig.TeamPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.ApplicantsPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.CalendarPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.ChatPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.CoursesPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.DocumentsPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.FinancesPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.JoinRequestsPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.MembersPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.MissionPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.PlanPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.ReportsPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.ResourcesPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.SurveyPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.TasksPage
+import org.ole.planet.myplanet.ui.teams.TeamPageConfig.TeamPage
 import org.ole.planet.myplanet.utilities.DialogUtils
 import org.ole.planet.myplanet.utilities.SharedPrefManager
 import org.ole.planet.myplanet.utilities.Utilities
