+++ app/src/main/java/org/ole/planet/myplanet/ui/teams/member/MemberFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.team.member
+package org.ole.planet.myplanet.ui.teams.member
 
 import android.content.res.Configuration
 import android.os.Bundle
