+++ app/src/main/java/org/ole/planet/myplanet/ui/teams/member/MembersViewModel.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.team.member
+package org.ole.planet.myplanet.ui.teams.member
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
